pyproject.toml
[tool.poetry]
name = "open_transit"
version = "0.0.1"
description = "A modern public transport information service"
authors = ["Dmitry Semenov <lonelyteapot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.3.0"
fastapi = "^0.103.1"
strawberry-graphql = { extras = ["fastapi"], version = "^0.205.0" }
uvicorn = "^0.23.2"
rich = "^13.5.2"
httpx = "^0.24.1"
jsonrpcclient = "^4.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.isort]
lines-after-imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
